[REFACTOR] common 디렉토리로 sendGetRequest 이동
각 디렉토리에서 공통으로 사용할 수 있게 함
